fr449

+++ project5/fractal_float_cuda.cu
@@ -41,7 +41,7 @@ static __global__ void  fractalKernel(const int width, const int frames, unsigne
   const long idx = threadIdx.x + blockIdx.x * (long)blockDim.x;
   if(idx < n){
   const int frame = idx / (width * width);
-    const float delta = Delta * pow(0.98f, frame);
+    const float delta = Delta * powf(0.98f, frame);
     const float xMin = xMid - delta;
     const float yMin = yMid - delta;
     const float dw = 2.0f * delta / width;
